--- new_task2_chp/templates/new_task2_chp/Results.html
@@ -1,12 +1,0 @@
-{% extends "global/Page.html" %}
-{% load otree static %}
-
-{% block title %}
-    Task 2: Results
-{% endblock %}
-
-{% block content %}
-    <p>Your points in this task are {{player.payoff}}.</p>
-
-    {% next_button %}
-{% endblock %}
