+++ settings.py
@@ -47,7 +47,7 @@ APP_KITS = {
             'new_task2_instructions',
             'new_task2_practice_chp',
             'new_task2_quiz',
-            'new_task2_chp',
+            'task2_tr2',
             'new_task2_results',
             'static_portfolio_chp',
             'static_portfolio_results',
@@ -105,7 +105,7 @@ SESSION_CONFIGS = [
                       'new_task2_practice_chp',
                       'new_task2_quiz',
                       'task2_tr1',
-                      'new_task2_chp',
+                      'task2_tr2',
                       'new_task2_results',
                       'static_portfolio',
                       'static_portfolio_chp',
