+++ task2_tr2/templates/task2_tr2/Results.html
@@ -1,0 +1,12 @@
+{% extends "global/Page.html" %}
+{% load otree static %}
+
+{% block title %}
+    Task 2: Results
+{% endblock %}
+
+{% block content %}
+    <p>Your points in this task are {{player.payoff}}.</p>
+
+    {% next_button %}
+{% endblock %}
